requirements.txt
django_tables2==2.0.3
django_filter==2.0.0
drf_tracking==1.5.0
Django==2.1.5
python_dateutil==2.7.5
django_filters==0.2.1
djangorestframework==3.9.1
